+++ gems/quilt_rails/lib/quilt_rails/railtie.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module Quilt
+  class Railtie < Rails::Railtie
+  end
+end