+++ packages/koa-shopify-auth/src/verify-request/verify-request.ts
@@ -1,5 +1,5 @@
 import {Context} from 'koa';
-
+import {Method, Header, StatusCode} from '@shopify/network';
 import {NextFunction} from '../types';
 import {TEST_COOKIE_NAME, TOP_LEVEL_OAUTH_COOKIE_NAME} from '../index';
 
@@ -20,9 +20,28 @@ export default function verifyRequest({
       query: {shop},
       session,
     } = ctx;
+    const redirect = `${authRoute}?shop=${shop}`;
 
     if (session && session.accessToken) {
       ctx.cookies.set(TOP_LEVEL_OAUTH_COOKIE_NAME);
+      // If a user has installed the store previously on their shop, the accessToken can be stored in session.
+      // we need to check if the accessToken is valid, and the only way to do this is by hitting the api.
+      const response = await fetch(
+        `https://${session.shop}/admin/metafields.json`,
+        {
+          method: Method.Post,
+          headers: {
+            [Header.ContentType]: 'application/json',
+            'X-Shopify-Access-Token': session.accessToken,
+          },
+        },
+      );
+
+      if (response.status === StatusCode.Unauthorized) {
+        ctx.redirect(redirect);
+        return;
+      }
+
       await next();
       return;
     }
@@ -30,7 +49,7 @@ export default function verifyRequest({
     ctx.cookies.set(TEST_COOKIE_NAME, '1');
 
     if (shop) {
-      ctx.redirect(`${authRoute}?shop=${shop}`);
+      ctx.redirect(redirect);
       return;
     }