+++ packages/polyfills/src/url.ts
@@ -1,1 +1,0 @@
-require('url-search-params-polyfill');