+++ packages/react-html/src/tests/serializer.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {extract} from '@shopify/react-effect/server';
-import {useSerialized, HtmlContext, HtmlManager} from '..';
 import {render, Html} from '../server';
+import {useSerialized, HtmlContext, HtmlManager} from '..';
 
 describe('useSerialized', () => {
   it('serializes promise results', async () => {