+++ packages/react-self-serializers/src/index.ts
@@ -1,0 +1,2 @@
+export {I18n} from './I18n';
+export {GraphQL} from './GraphQLComponent';