+++ packages/react-server-webpack-plugin/src/index.ts
@@ -1,0 +1,1 @@
+export {ReactServerPlugin} from './react-server-webpack-plugin';