+++ packages/react-testing/src/index.ts
@@ -1,4 +1,5 @@
 export {Root} from './root';
 export {Element} from './element';
+export {Node} from './types';
 export * from './mount';
 export * from './destroy';