+++ packages/react-testing/src/matchers/types.ts
@@ -1,6 +1,0 @@
-import {Root} from '../root';
-import {Element} from '../element';
-
-export {Root, Element};
-
-export type Node<Props> = Root<Props> | Element<Props>;