+++ packages/react-testing/src/types.ts
@@ -1,4 +1,13 @@
 import * as React from 'react';
+import {Arguments, MaybeFunctionReturnType} from '@shopify/useful-types';
+
+export type PropsFor<
+  T extends string | React.ComponentType<any>
+> = T extends string
+  ? React.HTMLAttributes<T>
+  : T extends React.ComponentType<any>
+    ? React.ComponentPropsWithoutRef<T>
+    : never;
 
 export type FunctionKeys<T> = {
   [K in keyof T]-?: NonNullable<T[K]> extends ((...args: any[]) => any)
@@ -60,4 +69,46 @@ export interface Fiber {
 
 export interface ReactInstance {
   _reactInternalFiber: Fiber;
+}
+
+export type Predicate = (node: Node<unknown>) => boolean;
+
+export interface Node<Props> {
+  readonly props: Props;
+  readonly type: string | React.ComponentType<any> | null;
+  readonly isDOM: boolean;
+  readonly instance: any;
+  readonly children: Node<unknown>[];
+  readonly descendants: Node<unknown>[];
+  readonly domNodes: HTMLElement[];
+  readonly domNode: HTMLElement | null;
+
+  data(key: string): string | undefined;
+  prop<K extends keyof Props>(key: K): Props[K];
+
+  text(): string;
+  html(): string;
+
+  is<Type extends React.ComponentType<any> | string>(
+    type: Type,
+  ): this is Node<PropsFor<Type>>;
+
+  find<Type extends React.ComponentType<any> | string>(
+    type: Type,
+    props?: Partial<PropsFor<Type>>,
+  ): Node<PropsFor<Type>> | null;
+  findAll<Type extends React.ComponentType<any> | string>(
+    type: Type,
+    props?: Partial<PropsFor<Type>>,
+  ): Node<PropsFor<Type>>[];
+  findWhere(predicate: Predicate): Node<unknown> | null;
+  findAllWhere(predicate: Predicate): Node<unknown>[];
+
+  trigger<K extends FunctionKeys<Props>>(
+    prop: K,
+    ...args: DeepPartialArguments<Arguments<Props[K]>>
+  ): MaybeFunctionReturnType<NonNullable<Props[K]>>;
+  triggerKeypath<T = unknown>(keypath: string, ...args: unknown[]): T;
+
+  toString(): string;
 }