+++ packages/with-env/src/types.ts
@@ -1,0 +1,1 @@
+export type Env = string | {[key: string]: string};