+++ packages/react-graphql/CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [4.0.4] - 2019-08-13
+
+### Fixed
+
+- Add missing dependencies ([832](https://github.com/Shopify/quilt/pull/832))
+
+## [4.0.3] - 2019-07-07
+
+### Added
+
+- Adds SSR extractable link to support GraphQL Self-serialization ([810](https://github.com/Shopify/quilt/pull/810))
 
 ## [4.0.2] - 2019-07-05